LayoutTests/fast/parser/comment-in-textarea-expected.txt

Test that the HTML parser correctly parses the comment-like string in the textarea.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textAreas.length is 1
PASS textAreas[0].innerHTML is "&lt;!-- &lt;/textarea&gt; --&gt; This should be part of the textarea"
PASS successfullyParsed is true

TEST COMPLETE
